[UDT-208] fix: 온보딩 빌드 오류 해결
